Move MissingIndicator to impute.py
