+++ sklearn/preprocessing/imputation.py
@@ -21,7 +21,7 @@ zip = six.moves.zip
 map = six.moves.map
 
 __all__ = [
-    'Imputer', 'MissingIndicator'
+    'Imputer',
 ]
 
 
@@ -371,201 +371,3 @@ class Imputer(BaseEstimator, TransformerMixin):
             X[coordinates] = values
 
         return X
-
-
-class MissingIndicator(BaseEstimator, TransformerMixin):
-    """Binary indicators for missing values.
-
-    Parameters
-    ----------
-    missing_values : integer or np.NaN, optional (default=np.NaN)
-        The placeholder for the missing values. All occurrences of
-        ``missing_values`` will be represented as ones.
-
-    features : {'auto' (default), 'all', array-like of int}
-        If "all", mask will represent all features.
-        If "auto", mask will only represent features with missing values
-        during fit time.
-        If mask/indices, mask will only represent features in the
-        indices or mask.
-
-    sparse : boolean or "auto", optional (default="auto")
-        If True, the transformed ``X`` will be a sparse matrix.
-        If False, the transformed ``X`` will be a numpy array.
-        If "auto", the transformed ``X`` will be of same type as input.
-
-    error_on_new : boolean, optional (default=True)
-        If True, transform will raise an error when there are features with
-        missing values in transform but have no missing values in fit
-        This is applicable only when ``features`` = "auto"
-
-    Example
-    -------
-    >>> from sklearn.preprocessing import MissingIndicator
-    >>> import numpy as np
-    >>> X1 = np.array([
-    ...       [np.nan,  1,  3],
-    ...       [ 4,  0, np.nan],
-    ...       [ 8,  1,  0]
-    ... ])
-    >>> X2 = np.array([
-    ...       [ 5, 1, np.nan],
-    ...       [np.nan,  2,  3],
-    ...       [ 2,  4,  0]
-    ... ])
-    >>> indicator = MissingIndicator()
-    >>> indicator.fit(X1)
-    MissingIndicator(error_on_new=True, features='auto', missing_values='NaN',
-             sparse='auto')
-    >>> X2_tr = indicator.transform(X2)
-    >>> X2_tr
-    array([[0, 1],
-           [1, 0],
-           [0, 0]])
-
-    Attributes
-    ----------
-    features_  : array of shape(n_missing_features,)
-        The features with missing values.
-
-    n_features_ : int
-        The number of features in the input.
-    """
-
-    def __init__(self, missing_values="NaN", features="auto", sparse="auto",
-                 error_on_new=True):
-        self.missing_values = missing_values
-        self.features = features
-        self.sparse = sparse
-        self.error_on_new = error_on_new
-
-    def fit(self, X, y=None):
-        """Fit the transformer on X.
-
-        Parameters
-        ----------
-        X : {array-like, sparse matrix}, shape (n_samples, n_features)
-            Input data, where ``n_samples`` is the number of samples and
-            ``n_features`` is the number of features.
-
-        Returns
-        -------
-        self : object
-            Returns self.
-        """
-        self._validate_params()
-
-        X = check_array(X, accept_sparse=('csc', 'csr'),
-                        force_all_finite=False)
-        self.n_features_ = X.shape[1]
-
-        if isinstance(self.features, six.string_types):
-            if self.features == "auto":
-                _, self.features_ = self._get_missing_features_info(X)
-            else:  # self.features == "all"
-                self.features_ = np.arange(self.n_features_)
-        else:
-            self.features_ = self.features
-
-        return self
-
-    def transform(self, X):
-        """Generate missing values indicator for X.
-
-        Parameters
-        ----------
-        X : {array-like, sparse matrix}, shape = [n_samples, n_features]
-            The input data to complete.
-
-        Returns
-        -------
-        Xt : array or sparse matrix, shape = [n_samples, n_features]
-             The missing indicator for input data
-
-        """
-        check_is_fitted(self, "features_", "n_features_")
-
-        X = check_array(X, accept_sparse=('csc', 'csr'),
-                        force_all_finite=False)
-        if X.shape[1] != self.n_features_:
-            raise ValueError("X has a different number of features "
-                             "than during fitting.")
-
-        imputer_mask, features_ = self._get_missing_features_info(X)
-
-        if isinstance(self.features, six.string_types):
-            if self.features == "auto":
-                features = np.setdiff1d(features_,
-                                        self.features_)
-                if self.error_on_new and features.size > 0:
-                    raise ValueError("The features %s have missing values "
-                                     "in transform but have no missing values "
-                                     "in fit" % features)
-
-        if not (isinstance(self.features, six.string_types) and
-                self.features == "all") and len(self.features_) != 0:
-            # no need to slice when all features have missing values
-            imputer_mask = imputer_mask[:, self.features_]
-
-        return imputer_mask
-
-    def fit_transform(self, X, y=None):
-        """Generate missing values indicator for X.
-
-        Parameters
-        ----------
-        X : {array-like, sparse matrix}, shape = [n_samples, n_features]
-            The input data to complete.
-
-        Returns
-        -------
-        Xt : array or sparse matrix, shape = [n_samples, n_features]
-             The missing indicator for input data
-
-        """
-        self.fit(X, y)
-        return self.transform(X)
-
-    def _validate_params(self):
-        if (isinstance(self.features, six.string_types) and
-                self.features not in ["auto", "all"]):
-            raise ValueError("Can only use these options: 'auto', 'all'"
-                             " got {0}".format(self.features))
-        elif not isinstance(self.features, six.string_types):
-            self.features = check_array(self.features, ensure_2d=False)
-            if (isinstance(self.features, np.ndarray) and
-                    self.features.dtype.kind != 'i'):
-                raise ValueError("Features should be an array of integers")
-
-        if not ((isinstance(self.sparse, six.string_types) and
-                self.sparse == "auto") or isinstance(self.sparse, bool)):
-            raise ValueError("sparse can only boolean or 'auto'"
-                             " got {0}".format(self.sparse))
-
-    def _get_missing_features_info(self, X):
-        if sparse.issparse(X) and self.missing_values != 0:
-            #  sparse matrix and missing values is not zero
-            imputer_mask = _get_mask(X.data, self.missing_values)
-            imputer_mask = X.__class__((imputer_mask, X.indices.copy(),
-                                        X.indptr.copy()), shape=X.shape,
-                                       dtype=X.dtype)
-            feat_with_missing = imputer_mask.sum(axis=0).nonzero()[1]
-        else:
-            #  sparse with zero as missing value and dense matrix
-            if sparse.issparse(X):
-                X = X.toarray()
-            imputer_mask = _get_mask(X, self.missing_values)
-            #  convert boolean mask to binary mask
-            imputer_mask = imputer_mask.astype(int, copy=False)
-            feat_with_missing = imputer_mask.sum(axis=0).nonzero()[0]
-
-        if ((self.sparse == 'auto' and sparse.issparse(imputer_mask)) or
-                self.sparse is True):
-            imputer_mask = sparse.csc_matrix(imputer_mask)
-        elif self.sparse is False and sparse.issparse(imputer_mask):
-            imputer_mask = imputer_mask.toarray()
-
-        if isinstance(feat_with_missing, np.matrix):
-            feat_with_missing = feat_with_missing.A1
-
-        return imputer_mask, feat_with_missing
